refactor : 코드래빗 PR 리뷰 보완

+++ apps/commerce-api/src/main/java/com/loopers/domain/common/vo/Money.java
@@ -32,7 +32,7 @@ public class Money {
         return new Money(new BigDecimal(amount));
     }
 
-    public static Money of(int amount) {
+    public static Money of(long amount) {
         return new Money(new BigDecimal(amount));
     }
